refactor: oracle bonded keeper deprecation
